legal-clause-extractor/requirements.txt
streamlit==1.35.0
pymupdf==1.24.5
google-generativeai==0.7.2
pandas==2.2.2
